+++ cookbooks/maxlab_base/spec/unit/recipes/default_spec.rb
@@ -20,11 +20,24 @@ describe 'maxlab_base::default' do
     end
   end
 
-  context 'When all attributes are default, on CentOS 7.4.1708' do
+  context 'When all attributes are default, on CentOS 8' do
     let(:chef_run) do
       # for a complete list of available platforms and versions see:
       # https://github.com/customink/fauxhai/blob/master/PLATFORMS.md
-      runner = ChefSpec::ServerRunner.new(platform: 'centos', version: '7.4.1708')
+      runner = ChefSpec::ServerRunner.new(platform: 'centos', version: '8')
+      runner.converge(described_recipe)
+    end
+
+    it 'converges successfully' do
+      expect { chef_run }.to_not raise_error
+    end
+  end
+
+  context 'When all attributes are default, on Redhat 8' do
+    let(:chef_run) do
+      # for a complete list of available platforms and versions see:
+      # https://github.com/customink/fauxhai/blob/master/PLATFORMS.md
+      runner = ChefSpec::ServerRunner.new(platform: 'redhat', version: '8')
       runner.converge(described_recipe)
     end